requirements.txt
Flask==2.3.2
flask-cors==3.0.10

numpy==1.26.4
opencv-python-headless==4.8.1.78
Pillow==10.0.0

pypdf==3.16.0
reportlab==4.4.9

pytesseract==0.3.10

torch>=2.1.0
torchvision>=0.16.0
python-doctr>=0.8.1
